requirements.txt
Flask
line-bot-sdk
requests
replicate
gunicorn==21.2.0
python-dotenv
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9